moved submit metric update to its own file
